Connected content to the server, however getting issues with authorization tokens on postman

+++ server/content/content.service.js
@@ -2,7 +2,10 @@ const config = require('../config.json');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
 const db = require('../_helpers/db');
+
 const Content = db.Content; 
+const User = db.User;
+
 module.exports = {
     getAll,
     getById,
@@ -10,12 +13,23 @@ module.exports = {
     update,
     delete: _delete
 };
+//FIX: Generic errors currently, might be helpful to have an error with a status code and a message? 
 
 //Returns all content
-async function getAll() {
-    return await Content.find();
+//FIX: Add if statements to check if it should be viewable? eg. viewPerms or check that the user can actually 
+//call this function 
+async function getAll(userid) {
+    try{ 
+        //Returns all content with the specified userid 
+        const posts = await Content.find({user: userid});
+    } catch(error) {
+        if (error.name == "CastError") { return null; }
+        throw error;
+    }
 }
 
+//Given a postid, gets the post
+//FIX: Again need to check post perms 
 async function getById(id) {
     try {
         return await Content.findById(id);
@@ -26,33 +40,62 @@ async function getById(id) {
     }
 }
 
-//userParam is the input recieved from the user (not related to the user.model I'm guessing)
-//Not sure if you need a try catch here, as you can create more than one post with the same content 
-async function create(userParam) {
-    //Create a new post 
-    const post = new Content(userParam); 
-    await post.save(); 
+//Creates a post
+async function create(userid, userParam) {
+    try {
+        //Create a new post 
+        const post = new Content(userParam); 
+        //Sets the user of the post to the person who created it 
+        post.user = userid; 
+        //Saves post and returns a status message and the post 
+        await post.save(); 
+        res.status(201).send(post.toObject());
+    } catch(error) {
+        //Otherwise throws error if post not created 
+        if (error.name == "CastError") { return null; }
+        throw error;
+    }
 }
 
-async function update(id, userParam) {
+//Updates a post
+async function update(userid, postid, userParam) {
     //attempt to get user, return null if error getting user
     let post;
     try {
-        post = await Content.findById(id);
+        post = await Content.findById(postid);
+        //Checks if the person editing is the the user of the post 
+        if(userid == post.user){ 
+            //Updates the specified post with the input 
+            Object.assign(post, userParam);
+            await post.save();
+            return post;
+        }
+        //Otherwise return an error (should I be throwing this?)
+        else{ 
+            res.status(400).send({
+                error: `You do not have permissions to edit this post`
+            });
+        }
     } catch(error) {
         //error thrown when not found in database
         if (error.name == "CastError") { return null; }
         throw error;
     }
-    //Updates the specified post with the input 
-    Object.assign(post, userParam);
-    await post.save();
-    return user;
 }
 
-async function _delete(id) {
+//Deletes a post
+async function _delete(userid, postid) {
     try {
-        return await Content.findByIdAndRemove(id);
+        post = Content.findbyId(postid); 
+        //Checks that the user accessing is the same as the creator 
+        if(userid == post.user){
+            return await Content.findByIdAndRemove(id);
+        }
+        else{ 
+            res.status(400).send({
+                error: `You do not have permissions to delete this post`
+            });
+        }
     } catch(error) {
         if (error.name == "CastError") { return null; }
         throw error;
